Setting up Jekyll
I'm going to use Jekyll as the blogging platform for this site. It is
the one recommended by GithubPages, and so I'm going to go wth it. This
is obviously a WIP, but I want to see what the result actually is.

I am, here, not "building" the project. That may be a mistake, but I
won't know until after I push it to master. Here goes nothing!